+++ deltachat-jsonrpc/src/api/types/message.rs
@@ -388,7 +388,7 @@ impl MessageSearchResult {
 }
 
 #[derive(Serialize, TypeDef)]
-#[serde(rename_all = "camelCase", rename = "MessageListItem")]
+#[serde(rename_all = "camelCase", rename = "MessageListItem", tag = "kind")]
 pub enum JSONRPCMessageListItem {
     Message {
         msg_id: u32,
@@ -397,7 +397,7 @@ pub enum JSONRPCMessageListItem {
     /// Day marker, separating messages that correspond to different
     /// days according to local time.
     DayMarker {
-        /// Marker timestamp, for day markers
+        /// Marker timestamp, for day markers, in unix milliseconds
         timestamp: i64,
     },
 }